ForLocalCommit
加入fragment.
骨架搭好。
