+++ app/src/main/java/com/example/beikeapp/TeacherRegister/TeacherRegister_FirstActivity.java
@@ -5,6 +5,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
+import android.support.v7.app.AppCompatActivity;
 import android.text.TextUtils;
 import android.util.Log;
 import android.view.View;
@@ -24,7 +25,7 @@ import java.util.List;
 import cn.smssdk.EventHandler;
 import cn.smssdk.SMSSDK;
 
-public class TeacherRegister_FirstActivity extends BaseActivity implements View.OnClickListener {
+public class TeacherRegister_FirstActivity extends AppCompatActivity implements View.OnClickListener {
 
     private FormEditText etPhone;
     private FormEditText etPsw;
