+++ app/src/main/java/com/example/beikeapp/TeacherRegister/TeacherRegister_SecondActivity.java
@@ -2,6 +2,7 @@ package com.example.beikeapp.TeacherRegister;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 
@@ -24,7 +25,7 @@ import java.util.List;
 
 import mabbas007.tagsedittext.TagsEditText;
 
-public class TeacherRegister_SecondActivity extends BaseActivity implements View.OnClickListener {
+public class TeacherRegister_SecondActivity extends AppCompatActivity implements View.OnClickListener {
     //上一活动传来的手机和密码
     private String phoneNumber, password;
     //带错误校验的输入框
