--- app/src/main/java/com/example/beikeapp/TeacherRegister/Teacher_MainActivity.java
@@ -1,15 +1,0 @@
-package com.example.beikeapp.TeacherRegister;
-
-import android.os.Bundle;
-
-import com.example.beikeapp.R;
-import com.example.beikeapp.Util.BaseActivity;
-
-public class Teacher_MainActivity extends BaseActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.teacher_main);
-    }
-}
