+++ app/src/main/java/com/example/beikeapp/TeacherWoDe/WoDeFragment.java
@@ -1,4 +1,4 @@
-package com.example.beikeapp.TeacherChat;
+package com.example.beikeapp.TeacherWoDe;
 
 
 import android.os.Bundle;
@@ -12,10 +12,10 @@ import com.example.beikeapp.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link FirstFragment#newInstance} factory method to
+ * Use the {@link WoDeFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class FirstFragment extends Fragment {
+public class WoDeFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -26,7 +26,7 @@ public class FirstFragment extends Fragment {
     private String mParam2;
 
 
-    public FirstFragment() {
+    public WoDeFragment() {
         // Required empty public constructor
     }
 
@@ -36,11 +36,11 @@ public class FirstFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment FirstFragment.
+     * @return A new instance of fragment WoDeFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static FirstFragment newInstance(String param1, String param2) {
-        FirstFragment fragment = new FirstFragment();
+    public static WoDeFragment newInstance(String param1, String param2) {
+        WoDeFragment fragment = new WoDeFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -61,7 +61,7 @@ public class FirstFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_first, container, false);
+        return inflater.inflate(R.layout.fragment_wode, container, false);
     }
 
 }
