+++ app/src/main/res/layout/fragment_shisheng.xml
@@ -7,6 +7,6 @@
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="ThirdFragment" />
+        android:text="ShiShengFragment" />
 
 </FrameLayout>
