--- app/src/main/res/layout/teacher_chat_classlist.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:hjm="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.example.beikeapp.TeacherChat.TeacherChatClassList">
-
-    <TextView
-        android:id="@+id/tv_classList"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:textSize="30dp"
-        android:text="班级列表"
-        tools:layout_editor_absoluteX="163dp"
-        tools:layout_editor_absoluteY="0dp" />
-
-    <ListView
-        android:id="@+id/lv_classList"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
-        android:layout_below="@+id/tv_classList"/>
-
-    <com.hjm.bottomtabbar.BottomTabBar
-        android:id="@+id/bottom_tab_bar"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        hjm:tab_bar_background="#FFFFFF"
-        hjm:tab_divider_background="#FF0000"
-        hjm:tab_font_size="14px"
-        hjm:tab_img_height="70px"
-        hjm:tab_img_width="70px"
-        hjm:tab_isshow_divider="true"
-        hjm:tab_padding_bottom="5px"
-        hjm:tab_padding_top="5px"
-        hjm:tab_selected_color="#000000"
-        hjm:tab_unselected_color="@color/colorPrimary"/>
-</RelativeLayout>
