+++ app/src/main/res/layout/teacher_main_function.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:hjm="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <!--<fragment-->
+        <!--android:layout_width="wrap_content"-->
+        <!--android:layout_height="wrap_content"-->
+        <!--android:id="@+id/fragment_shisheng"-->
+        <!--class="com.example.beikeapp.TeacherShiSheng.ShiShengFragment"/>-->
+
+    <!--<fragment-->
+        <!--android:layout_width="wrap_content"-->
+        <!--android:layout_height="wrap_content"-->
+        <!--android:id="@+id/fragment_jiaxiao"-->
+        <!--class="com.example.beikeapp.TeacherJiaXiao.JiaXiaoFragment"/>-->
+
+    <!--<fragment-->
+        <!--android:layout_width="wrap_content"-->
+        <!--android:layout_height="wrap_content"-->
+        <!--android:id="@+id/fragment_wode"-->
+        <!--class="com.example.beikeapp.TeacherWoDe.WoDeFragment"/>-->
+
+    <com.hjm.bottomtabbar.BottomTabBar
+        android:id="@+id/bottom_tab_bar"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        hjm:tab_bar_background="#FFFFFF"
+        hjm:tab_divider_background="#5F00FF"
+        hjm:tab_font_size="14px"
+        hjm:tab_img_height="70px"
+        hjm:tab_img_width="70px"
+        hjm:tab_isshow_divider="true"
+        hjm:tab_padding_bottom="5px"
+        hjm:tab_padding_top="5px"
+        hjm:tab_selected_color="#000000"
+        hjm:tab_unselected_color="@color/colorPrimary"/>
+</RelativeLayout>
